% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altw.R
\name{altw}
\alias{altw}
\title{Use clipr to copy selection(s) in current source editor to X clipboard}
\usage{
altw()
}
\description{
Use clipr to copy selection(s) in current source editor to X clipboard
}
\author{
fnaufel
}
