% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xclip.R
\name{xclip}
\alias{xclip}
\title{Use xclip to copy selection(s) in current source editor to X clipboard}
\usage{
xclip()
}
\description{
Use xclip to copy selection(s) in current source editor to X clipboard
}
\author{
fnaufel
}
